requirements.txt
fastapi
uvicorn
openai
faiss-cpu
pydantic
requests
numpy
standard-imghdr
python-dotenv